pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "autocertsync"
version = "1.0.0"
description = "自动 SSL/TLS 证书同步工具"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "paramiko>=3.4.0",
    "scp>=0.14.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0",
    "websockets>=12.0",
]

[project.scripts]
autocertsync = "autocertsync.__main__:main"
